DEV-184 - Added rpc-client profile for plugin service
