+++ devicehive-plugin/src/main/java/com/devicehive/service/KafkaWsTopicService.java
@@ -1,0 +1,46 @@
+package com.devicehive.service;
+
+/*
+ * #%L
+ * DeviceHive Frontend Logic
+ * %%
+ * Copyright (C) 2016 - 2017 DataArt
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+
+import com.devicehive.proxy.PluginProxyClient;
+import com.devicehive.shim.kafka.topic.KafkaTopicService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Profile;
+import org.springframework.stereotype.Component;
+
+import java.util.Arrays;
+
+@Component
+@Profile("ws-kafka-proxy")
+public class KafkaWsTopicService implements KafkaTopicService {
+
+    private final PluginProxyClient pluginProxyClient;
+    
+    @Autowired
+    public KafkaWsTopicService(PluginProxyClient pluginProxyClient) {
+        this.pluginProxyClient = pluginProxyClient;
+    }
+
+    @Override
+    public void createTopic(String topic) {
+        pluginProxyClient.createTopic(Arrays.asList(topic));
+    }
+}
