+++ devicehive-plugin/src/main/resources/application.properties
@@ -19,7 +19,7 @@
 ###
 server.port = 8110
 
-spring.profiles.active=ws-kafka-proxy
+spring.profiles.active=rpc-client
 
 logging.level.org.springframework.security=INFO
