+++ dockerfiles/devicehive-frontend-rdbms/devicehive-start.sh
@@ -89,6 +89,6 @@ java -server -Xms1g -Xmx2g -XX:+UseG1GC -XX:MaxGCPauseMillis=20 -XX:+DisableExpl
 -Dzookeeper.connect="${DH_ZK_ADDRESS}:${DH_ZK_PORT:-2181}" \
 -Dzookeeper.connectionTimeout="${DH_ZK_CONNECTIONTIMEOUT:-8000}" \
 -Dzookeeper.sessionTimeout="${DH_ZK_SESSIONTIMEOUT:-10000}" \
-"./devicehive-frontend-${DH_VERSION}-boot.jar" --spring.profiles.active=ws-kafka-proxy &
+"./devicehive-frontend-${DH_VERSION}-boot.jar" &
 PID=$!
 wait "$PID"
